replace cli
